Add shared useHexColormap to ProjectContext
